+++ vadl/main/vadl/pass/PassOrders.java
@@ -16,867 +16,53 @@
 
 package vadl.pass;
 
-import static vadl.configuration.DecoderOptions.Generator.RTL_TABLE;
-import static vadl.configuration.DecoderOptions.OptionalStep.OPT_ALL;
-import static vadl.configuration.DecoderOptions.OptionalStep.OPT_CONSTRAINT_SYNTHESIS;
-import static vadl.configuration.DecoderOptions.OptionalStep.OPT_DECODER_VERIFICATION;
-import static vadl.configuration.DecoderOptions.OptionalStep.OPT_ENCODING_VERIFICATION;
-import static vadl.iss.template.IssDefaultRenderingPass.issDefault;
-
-import com.google.common.collect.Streams;
 import java.io.IOException;
-import java.util.Optional;
-import java.util.stream.Stream;
 import javax.annotation.Nullable;
-import vadl.configuration.DumpMode;
 import vadl.configuration.GcbConfiguration;
 import vadl.configuration.GeneralConfiguration;
 import vadl.configuration.IssConfiguration;
 import vadl.configuration.LcbConfiguration;
 import vadl.configuration.RtlConfiguration;
-import vadl.dump.CollectBehaviorDotGraphPass;
-import vadl.dump.DumpIssInstructionGraphsPass;
-import vadl.dump.HtmlDumpPass;
-import vadl.gcb.passes.DetermineBuiltinAttributesPass;
-import vadl.gcb.passes.DetermineRegisterUsesAndDefsPass;
-import vadl.gcb.passes.DetermineRelocationTypeForFieldPass;
-import vadl.gcb.passes.GenerateCompilerRegistersPass;
-import vadl.gcb.passes.GenerateGcbIntrinsicsPass;
-import vadl.gcb.passes.GenerateValueRangeImmediatePass;
-import vadl.gcb.passes.IdentifyFieldUsagePass;
-import vadl.gcb.passes.InstructionPatternPruningPass;
-import vadl.gcb.passes.PredicateFunctionInlinerPass;
-import vadl.gcb.passes.RenamingConflictingRegistersPass;
-import vadl.gcb.passes.SetMissingConfigurationValuesPass;
-import vadl.gcb.passes.assembly.AssemblyConcatBuiltinMergingPass;
-import vadl.gcb.passes.encodingGeneration.GenerateFieldAccessEncodingAndPredicateFunctionsPass;
-import vadl.gcb.passes.operands.GenerateInstructionOperandsPass;
-import vadl.iss.passes.IssApplyMemoryEndiannessPass;
-import vadl.iss.passes.IssBitfieldWriteLoweringPass;
-import vadl.iss.passes.IssBuiltInArgTruncOptPass;
-import vadl.iss.passes.IssCFunctionExtractionPass;
-import vadl.iss.passes.IssCommonExprSavePass;
-import vadl.iss.passes.IssConfigurationPass;
-import vadl.iss.passes.IssExecStrategyPass;
-import vadl.iss.passes.IssExtractOptimizationPass;
-import vadl.iss.passes.IssGdbInfoExtractionPass;
-import vadl.iss.passes.IssHardcodedTcgAddOnPass;
-import vadl.iss.passes.IssInfoRetrievalPass;
-import vadl.iss.passes.IssLoopUnrollPass;
-import vadl.iss.passes.IssMemoryAccessTransformationPass;
-import vadl.iss.passes.IssMemoryDetectionPass;
-import vadl.iss.passes.IssNormalizationPass;
-import vadl.iss.passes.IssPcAccessConversionPass;
-import vadl.iss.passes.IssRegisterAccessInfoRetrievalPass;
-import vadl.iss.passes.IssRegisterAccessLoweringPass;
-import vadl.iss.passes.IssSelectLoweringPass;
-import vadl.iss.passes.IssTcgSchedulingPass;
-import vadl.iss.passes.IssTcgVAllocationPass;
-import vadl.iss.passes.IssTensorAssignmentToForallPass;
-import vadl.iss.passes.opDecomposition.IssOpDecompositionPass;
-import vadl.iss.passes.safeResourceRead.IssSafeResourceReadPass;
-import vadl.iss.passes.tcgLowering.IssTcgContextPass;
-import vadl.iss.passes.tcgLowering.TcgBranchLoweringPass;
-import vadl.iss.passes.tcgLowering.TcgOpLoweringPass;
-import vadl.iss.template.gdb_xml.EmitIssGdbXmlPass;
-import vadl.iss.template.hw.EmitIssHwMachineCPass;
-import vadl.iss.template.target.EmitIssCpuHeaderPass;
-import vadl.iss.template.target.EmitIssCpuParamHeaderPass;
-import vadl.iss.template.target.EmitIssCpuQomHeaderPass;
-import vadl.iss.template.target.EmitIssCpuSourcePass;
-import vadl.iss.template.target.EmitIssDecodeTreePass;
-import vadl.iss.template.target.EmitIssDoExcCIncPass;
-import vadl.iss.template.target.EmitIssGdbStubPass;
-import vadl.iss.template.target.EmitIssHelperCPass;
-import vadl.iss.template.target.EmitIssHelperHPass;
-import vadl.iss.template.target.EmitIssInsnTransCIncPass;
-import vadl.iss.template.target.EmitIssMachinePass;
-import vadl.iss.template.target.EmitIssTranslateCPass;
-import vadl.lcb.codegen.assembly.WrapInIntegralPass;
-import vadl.lcb.include.llvm.IR.EmitMiddleendIntrinsicsTableGenPass;
-import vadl.lcb.passes.OverwriteInputOperandsPass;
-import vadl.lcb.passes.asm.AsmGrammarRuleGenerationPass;
-import vadl.lcb.passes.isaMatching.IsaMachineInstructionMatchingPass;
-import vadl.lcb.passes.isaMatching.IsaPseudoInstructionMatchingPass;
-import vadl.lcb.passes.llvmLowering.CreateFunctionsFromImmediatesPass;
-import vadl.lcb.passes.llvmLowering.GenerateTableGenAbiSequenceInstructionRecordPass;
-import vadl.lcb.passes.llvmLowering.GenerateTableGenMachineInstructionRecordPass;
-import vadl.lcb.passes.llvmLowering.GenerateTableGenPseudoInstructionRecordPass;
-import vadl.lcb.passes.llvmLowering.GenerateTableGenRegistersPass;
-import vadl.lcb.passes.llvmLowering.ISelLoweringOperationActionPass;
-import vadl.lcb.passes.llvmLowering.LlvmLoweringPass;
-import vadl.lcb.passes.llvmLowering.RemoveRegisterWritesPass;
-import vadl.lcb.passes.llvmLowering.RemoveTruncationAndAnyExtPass;
-import vadl.lcb.passes.llvmLowering.compensation.CompensationPatternPass;
-import vadl.lcb.passes.llvmLowering.immediates.GenerateTableGenImmediateRecordPass;
-import vadl.lcb.passes.pseudo.AbiSequencesCompilerInstructionExpansionFunctionGeneratorPass;
-import vadl.lcb.passes.pseudo.PseudoExpansionFunctionGeneratorPass;
-import vadl.lcb.passes.relocation.GenerateLinkerComponentsPass;
-import vadl.lcb.template.clang.lib.Basic.Targets.EmitClangTargetHeaderFilePass;
-import vadl.lcb.template.clang.lib.CodeGen.EmitCGBuiltinFilePass;
-import vadl.lcb.template.clang.lib.CodeGen.EmitCodeGenFunctionHeaderFilePass;
-import vadl.lcb.template.include.llvm.BinaryFormat.ELFRelocs.EmitTargetElfRelocsDefFilePass;
-import vadl.lcb.template.lib.IR.EmitMiddleendMainIntrinsicsTableGenPass;
-import vadl.lcb.template.lib.Target.EmitMCInstLowerCppFilePass;
-import vadl.lcb.template.lib.Target.EmitMCInstLowerHeaderFilePass;
-import vadl.lcb.template.lib.Target.EmitVadlBuiltinHeaderFilePass;
-import vadl.lcb.template.lib.Target.EmitVadlBuiltinHeaderPPFilePass;
-import vadl.lcb.template.lib.Target.MCTargetDesc.EmitConstMatIntCppFilePass;
-import vadl.lcb.template.lib.Target.MCTargetDesc.EmitConstMatIntHeaderFilePass;
-import vadl.lcb.template.lib.Target.MCTargetDesc.EmitInstPrinterCppFilePass;
-import vadl.lcb.template.lib.Target.MCTargetDesc.EmitInstPrinterHeaderFilePass;
-import vadl.lcb.template.lld.ELF.Arch.EmitLldVadlBuiltinsHeaderFilePass;
-import vadl.rtl.passes.CleanupEmitDirectoryPass;
-import vadl.rtl.passes.ControlLogicPass;
-import vadl.rtl.passes.DebugOutputPass;
-import vadl.rtl.passes.EmitBuildSbtPass;
-import vadl.rtl.passes.EmitCoreEmitPass;
-import vadl.rtl.passes.EmitCoreTestPass;
-import vadl.rtl.passes.EmitElfSimPass;
-import vadl.rtl.passes.EmitModulesPass;
-import vadl.rtl.passes.EmitRVFIOutputsPass;
-import vadl.rtl.passes.EmitRtlDevcontainerConfigPass;
-import vadl.rtl.passes.EmitRtlDevcontainerDockerComposePass;
-import vadl.rtl.passes.EmitRtlMakefilePass;
-import vadl.rtl.passes.EmitScalafmtConfigPass;
-import vadl.rtl.passes.EmitSimMemCppPass;
-import vadl.rtl.passes.EmitSimMemSvPass;
-import vadl.rtl.passes.EmitVadlLibPass;
-import vadl.rtl.passes.ForwardingLogicPass;
-import vadl.rtl.passes.HazardAnalysisPass;
-import vadl.rtl.passes.InstructionProgressGraphCreationPass;
-import vadl.rtl.passes.InstructionProgressGraphLowerPass;
-import vadl.rtl.passes.InstructionProgressGraphMergePass;
-import vadl.rtl.passes.InstructionProgressGraphNamePass;
-import vadl.rtl.passes.MiaMappingCreationPass;
-import vadl.rtl.passes.MiaMappingInlinePass;
-import vadl.rtl.passes.MiaMappingOptimizePass;
-import vadl.rtl.passes.RtlConfigurationPass;
-import vadl.rtl.passes.StageOrderingPass;
-import vadl.template.AbstractTemplateRenderingPass;
-import vadl.vdt.passes.VdtConstraintSynthesisPass;
-import vadl.vdt.passes.VdtEncodingConstraintValidationPass;
-import vadl.vdt.passes.VdtEncodingSemanticVerificationPass;
-import vadl.vdt.passes.VdtInputPreparationPass;
-import vadl.vdt.passes.VdtLoweringPass;
-import vadl.vdt.passes.VdtVerificationPass;
+import vadl.pass.order.CheckPassOrder;
+import vadl.pass.order.GcbPassOrder;
+import vadl.pass.order.IssPassOrder;
+import vadl.pass.order.LcbPassOrder;
+import vadl.pass.order.RtlPassOrder;
+import vadl.pass.order.ViamPassOrder;
 import vadl.viam.Specification;
-import vadl.viam.passes.ControlFlowOptimizationPass;
-import vadl.viam.passes.DetectRegisterIndicesPass;
-import vadl.viam.passes.DuplicateWriteDetectionPass;
-import vadl.viam.passes.HardcodeLGALabelPass;
-import vadl.viam.passes.InstructionResourceAccessAnalysisPass;
-import vadl.viam.passes.NormalizeFieldsToFieldAccessFunctionsPass;
-import vadl.viam.passes.SnapshotInstructionBehaviorPass;
-import vadl.viam.passes.algebraic_simplication.AlgebraicSimplificationPass;
-import vadl.viam.passes.behaviorRewrite.BehaviorRewritePass;
-import vadl.viam.passes.canonicalization.CanonicalizationPass;
-import vadl.viam.passes.functionInliner.ArtificialResInlinerPass;
-import vadl.viam.passes.functionInliner.FieldAccessInlinerPass;
-import vadl.viam.passes.functionInliner.FunctionInlinerPass;
-import vadl.viam.passes.sideEffectScheduling.SideEffectSchedulingPass;
-import vadl.viam.passes.sideeffect_condition.SideEffectConditionResolvingPass;
-import vadl.viam.passes.staticCounterAccess.StaticCounterAccessResolvingPass;
-import vadl.viam.passes.statusBuiltInInlinePass.RemoveUnusedStatusFlagsFromBuiltinsPass;
-import vadl.viam.passes.statusBuiltInInlinePass.StatusBuiltInInlinePass;
-import vadl.viam.passes.verification.ViamVerificationPass;
 
 /**
- * This class contains static methods that define the individual pass orders for different
- * generation targets (e.g., LCB, ISS, ...).
+ * Compatibility facade for pass-order construction.
  */
 public class PassOrders {
-
-  /**
-   * Used by the {@code check} command.
-   * It doesn't apply transformation to the VIAM, however, it checks if the VDT can be constructed.
-   */
   public static PassOrder check(GeneralConfiguration configuration) {
-    var order = new PassOrder();
-    order.add(new ViamCreationPass(configuration));
-
-    addHtmlDump(order, configuration, "VIAM Creation",
-        "Dump directly after frontend generated VIAM.");
-
-    order.add(new ViamVerificationPass(configuration));
-
-    // check if VDT can be constructed
-    addDecodePasses(order, configuration);
-
-    addHtmlDump(order, configuration,
-        "VDT Creation",
-        "Dump directly after VDT generation.");
-
-    return order;
+    return CheckPassOrder.create(configuration);
   }
 
-  /**
-   * Return the viam passes.
-   */
   public static PassOrder viam(GeneralConfiguration configuration) throws IOException {
-    var order = new PassOrder();
-
-    // this is just a pseudo pass to add the behavior to the HTML dump
-    // at the stage directly after the VIAM creation.
-    order.add(new ViamCreationPass(configuration));
-    order.add(new ViamVerificationPass(configuration));
-
-    order.add(new DetectRegisterIndicesPass(configuration));
-    order.add(new OverwriteInputOperandsPass(configuration));
-    order.add(new NormalizeFieldsToFieldAccessFunctionsPass(configuration));
-    order.add(new RenamingConflictingRegistersPass(configuration));
-    order.add(new SnapshotInstructionBehaviorPass(configuration));
-
-    order.add(new RemoveUnusedStatusFlagsFromBuiltinsPass(configuration));
-    order.add(new StatusBuiltInInlinePass(configuration));
-
-    // Common optimizations
-    order.add(new CanonicalizationPass(configuration));
-    order.add(new AlgebraicSimplificationPass(configuration));
-    order.add(new BehaviorRewritePass(configuration));
-
-    // TODO: @kper do you see any fix for this?
-    // Note: we run the counter-access resolving pass before the func inliner pass
-    // because the lcb uses the uninlined version of the instructions.
-    // However, this might miss a lot of opportunities to statically resolve counter-accesses
-    // as the canonicalization runs at a later point.
-    order.add(new StaticCounterAccessResolvingPass(configuration));
-    order.add(new FunctionInlinerPass(configuration));
-    order.add(new FieldAccessInlinerPass(configuration));
-    order.add(new ArtificialResInlinerPass(configuration));
-    order.add(new ControlFlowOptimizationPass(configuration));
-    order.add(new SideEffectConditionResolvingPass(configuration));
-    // requires SideEffectConditionResolvingPass to work
-    order.add(new DuplicateWriteDetectionPass(configuration));
-
-    order.add(new CanonicalizationPass(configuration));
-    order.add(new AlgebraicSimplificationPass(configuration));
-    order.add(new BehaviorRewritePass(configuration));
-    order.add(new InstructionResourceAccessAnalysisPass(configuration));
-
-    // Hardcoded
-    order.add(new HardcodeLGALabelPass(configuration));
-
-    // verification after viam optimizations
-    order.add(new ViamVerificationPass(configuration));
-
-    return order;
+    return ViamPassOrder.create(configuration);
   }
 
-  /**
-   * Return the gcb and cppcodegen passes.
-   */
-  public static PassOrder gcbAndCppCodeGen(GcbConfiguration gcbConfiguration) throws IOException {
-    var order = viam(gcbConfiguration);
-    order.add(new SetMissingConfigurationValuesPass(gcbConfiguration));
-
-    // skip status built-in inlining as lcb
-    // needs to know about those status built-in calls.
-    order.skip(StatusBuiltInInlinePass.class);
-
-    // We need to keep the aliases.
-    order.skip(ArtificialResInlinerPass.class);
-
-
-    order.add(new GenerateCompilerRegistersPass(gcbConfiguration));
-    // skip inlining of field access
-    order.skip(FieldAccessInlinerPass.class);
-    order.add(new IdentifyFieldUsagePass(gcbConfiguration));
-    order.add(new DetermineRelocationTypeForFieldPass(gcbConfiguration));
-    order.add(new GenerateValueRangeImmediatePass(gcbConfiguration));
-    order.add(new GenerateFieldAccessEncodingAndPredicateFunctionsPass(gcbConfiguration));
-    order.add(new PredicateFunctionInlinerPass(gcbConfiguration));
-    order.add(new AssemblyConcatBuiltinMergingPass(gcbConfiguration));
-    order.add(new DetermineRegisterUsesAndDefsPass(gcbConfiguration));
-    order.add(new GenerateInstructionOperandsPass(gcbConfiguration));
-    order.add(new InstructionPatternPruningPass(gcbConfiguration));
-    order.add(new DetermineBuiltinAttributesPass(gcbConfiguration));
-    order.add(new GenerateGcbIntrinsicsPass(gcbConfiguration));
-
-    addHtmlDump(order, gcbConfiguration, "gcbProcessing",
-        "Now the gcb produced all necessary encoding function for field accesses "
-            + "and normalized VIAM types to Cpp types.");
-
-    return order;
+  public static PassOrder gcbAndCppCodeGen(GcbConfiguration configuration) throws IOException {
+    return GcbPassOrder.create(configuration);
   }
 
-  /**
-   * This is the pass order which must be executed to get a LLVM compiler.
-   */
-  public static PassOrder lcb(LcbConfiguration configuration)
-      throws IOException {
-    var order = gcbAndCppCodeGen(configuration);
-    // skip inlining of field access
-    order.skip(FieldAccessInlinerPass.class);
-
-    order.add(new PseudoExpansionFunctionGeneratorPass(configuration));
-    order.add(new AbiSequencesCompilerInstructionExpansionFunctionGeneratorPass(
-        configuration));
-
-    order.add(new RemoveTruncationAndAnyExtPass(configuration));
-
-    order.add(new GenerateTableGenRegistersPass(configuration));
-    order.add(new RemoveRegisterWritesPass(configuration));
-    order.add(new RemoveUnusedStatusFlagsFromBuiltinsPass(configuration));
-
-    // Common optimizations after register elimination
-    order.add(new CanonicalizationPass(configuration));
-    order.add(new AlgebraicSimplificationPass(configuration));
-    order.add(new BehaviorRewritePass(configuration));
-
-    order.add(new IsaMachineInstructionMatchingPass(configuration));
-    order.add(new IsaPseudoInstructionMatchingPass(configuration));
-
-    order.add(new LlvmLoweringPass(configuration));
-    order.add(new GenerateTableGenMachineInstructionRecordPass(configuration));
-    order.add(new GenerateTableGenPseudoInstructionRecordPass(configuration));
-    order.add(new GenerateTableGenAbiSequenceInstructionRecordPass(configuration));
-    order.add(new GenerateTableGenImmediateRecordPass(configuration));
-    order.add(new CreateFunctionsFromImmediatesPass(configuration));
-    order.add(new CompensationPatternPass(configuration));
-    order.add(new ISelLoweringOperationActionPass(configuration));
-    order.add(new GenerateLinkerComponentsPass(configuration));
-
-    order.add(new WrapInIntegralPass(configuration));
-
-    order.add(new AsmGrammarRuleGenerationPass(configuration));
-
-    addHtmlDump(order, configuration,
-        "lcbLlvmLowering",
-        "The LCB did ISA matching to and lowered common VIAM nodes to LLVM specific"
-            + "nodes.");
-
-    order.add(new EmitVadlBuiltinHeaderFilePass(configuration));
-    order.add(new EmitVadlBuiltinHeaderPPFilePass(configuration));
-    order.add(new vadl.lcb.template.clang.include.Basic.EmitCMakeListsPass(configuration));
-    order.add(new vadl.lcb.template.clang.include.Basic.EmitBuiltinsTableGenPass(configuration));
-    order.add(
-        new vadl.lcb.template.clang.include.Basic.EmitTargetBuiltinsHeaderPass(configuration));
-    order.add(new vadl.lcb.clang.lib.Driver.ToolChains.EmitClangToolChainFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.clang.lib.Driver.ToolChains.EmitClangCommonArgsToolChainFilePass(
-            configuration));
-    order.add(new EmitClangTargetHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.clang.lib.Basic.Targets.EmitClangTargetsFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.clang.lib.Basic.Targets.EmitClangTargetCppFilePass(configuration));
-    order.add(new vadl.lcb.template.clang.lib.Basic.EmitClangBasicCMakeFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.clang.lib.CodeGen.EmitCodeGenModuleCMakeFilePass(configuration));
-    order.add(new vadl.lcb.template.clang.lib.CodeGen.Targets.EmitClangCodeGenTargetFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.clang.lib.CodeGen.EmitCodeGenTargetInfoHeaderFilePass(configuration));
-    order.add(new vadl.lcb.clang.lib.CodeGen.EmitCodeGenModuleFilePass(configuration));
-    order.add(new vadl.lcb.template.lld.ELF.EmitLldDriverFilePass(configuration));
-    order.add(new vadl.lcb.template.lld.ELF.EmitLldELFCMakeFilePass(configuration));
-    order.add(new vadl.lcb.template.lld.ELF.EmitLldTargetHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lld.ELF.Arch.EmitLldTargetRelocationsHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lld.ELF.Arch.EmitLldManualEncodingHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lld.ELF.Arch.EmitImmediateUtilsHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lld.ELF.Arch.EmitLldArchFilePass(configuration));
-    order.add(new EmitLldVadlBuiltinsHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lld.ELF.EmitLldTargetCppFilePass(configuration));
-    order.add(new vadl.lcb.template.EmitLcbMakeFilePass(configuration));
-    order.add(new EmitTargetElfRelocsDefFilePass(
-        configuration));
-    order.add(new vadl.lcb.include.llvm.IR.EmitCMakeListsPass(configuration));
-    order.add(new EmitMiddleendMainIntrinsicsTableGenPass(configuration));
-    order.add(new EmitMiddleendIntrinsicsTableGenPass(configuration));
-    order.add(new EmitCGBuiltinFilePass(configuration));
-    order.add(new EmitCodeGenFunctionHeaderFilePass(configuration));
-    order.add(new vadl.lcb.include.llvm.BinaryFormat.EmitElfHeaderFilePass(configuration));
-    order.add(new vadl.lcb.include.llvm.Object.EmitELFObjectHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Misc.EmitBenchmarkRegisterHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitFrameLoweringCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.EmitMachineFunctionInfoHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetObjectFileCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitInstrInfoHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitDAGToDAGIselHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.AsmParser.EmitAsmParsedOperandCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.AsmParser.EmitAsmParsedOperandHeaderFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.AsmParser.EmitAsmRecursiveDescentParserHeaderFilePass(
-            configuration));
-    order.add(new vadl.lcb.template.lib.Target.AsmParser.EmitAsmParserCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.AsmParser.EmitAsmParserCMakeFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.AsmParser.EmitAsmRecursiveDescentParserCppFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.EmitDAGToDAGISelCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitAsmPrinterHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitCallingConvTableGenFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitRegisterInfoHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.Utils.EmitBaseInfoFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.Utils.EmitImmediateFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetTableGenFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitAsmPrinterCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitSubTargetHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitFrameLoweringHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitPassConfigHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitISelLoweringCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.TargetInfo.EmitTargetInfoHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.TargetInfo.EmitTargetInfoCMakeFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.TargetInfo.EmitTargetInfoCppFile(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitPassConfigCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitSubTargetCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetCMakeFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCCodeEmitterHeaderFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMcTargetDescCMakeFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCCodeEmitterCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitAsmStreamerCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitELFStreamerCppFilePass(configuration));
-    order.add(
-        new EmitConstMatIntHeaderFilePass(configuration));
-    order.add(
-        new EmitConstMatIntCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCInstExpanderCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitAsmBackendHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitELFObjectWriterCppFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCExprHeaderFilePass(configuration));
-    order.add(new EmitMCInstLowerCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCExprCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCTargetDescHeaderFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitAsmUtilsCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCTargetDescCppFilePass(configuration));
-    order.add(new EmitInstPrinterHeaderFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitAsmBackendCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCAsmInfoCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitELFStreamerHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitAsmStreamerHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitTargetStreamerHeaderFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCAsmInfoHeaderFilePass(configuration));
-    order.add(new EmitMCInstLowerHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitELFObjectWriterHeaderFilePass(
-        configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitFixupKindsHeaderFilePass(configuration));
-    order.add(
-        new EmitInstPrinterCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.MCTargetDesc.EmitAsmUtilsHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.MCTargetDesc.EmitMCInstExpanderHeaderFilePass(
-        configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitRegisterInfoTableGenFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitInstrInfoCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitInstrInfoTableGenFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitRegisterInfoCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetMachineCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetMachineHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetObjectFileHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitISelLoweringHeaderFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.lib.Target.EmitTargetTransformInfoHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Target.EmitTargetTransformInfoCppFilePass(configuration));
-    order.add(
-        new vadl.lcb.template.include.llvm.TargetParser.EmitTripleHeaderFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.TargetParser.EmitTripleCppFilePass(configuration));
-    order.add(new vadl.lcb.template.lib.Object.EmitElfCppFilePass(configuration));
-
-    return order;
+  public static PassOrder lcb(LcbConfiguration configuration) throws IOException {
+    return LcbPassOrder.create(configuration);
   }
 
-  /**
-   * Constructs the pass order used to generate the ISS (QEMU) from a VADL specification.
-   */
-  public static PassOrder iss(IssConfiguration config) throws IOException {
-    var order = viam(config);
-
-    // skip inlining of field access
-    order.skip(ArtificialResInlinerPass.class);
-    order.skip(FieldAccessInlinerPass.class);
-    order.skip(NormalizeFieldsToFieldAccessFunctionsPass.class);
-    order.skip(RenamingConflictingRegistersPass.class);
-    order.skip(OverwriteInputOperandsPass.class);
-
-    // iss function passes
-    order
-        .add(new IssInfoRetrievalPass(config))
-        .add(new IssConfigurationPass(config))
-        .add(new IssApplyMemoryEndiannessPass(config))
-        .add(new IssMemoryDetectionPass(config))
-        .add(new IssRegisterAccessLoweringPass(config))
-        .add(new IssExecStrategyPass(config))
-        .add(new IssBitfieldWriteLoweringPass(config))
-        .add(new IssTensorAssignmentToForallPass(config))
-        // run canonicalization, as register access lowering may create constant nodes
-        .add(new CanonicalizationPass(config))
-        .add(new IssOpDecompositionPass(config))
-        .add(new IssNormalizationPass(config))
-        .add(new IssExtractOptimizationPass(config))
-        .add(new IssMemoryAccessTransformationPass(config))
-        .add(new IssBuiltInArgTruncOptPass(config))
-        .add(new IssLoopUnrollPass(config))
-        .add(new SideEffectSchedulingPass(config))
-        .add(new IssSafeResourceReadPass(config))
-        .add(new IssCommonExprSavePass(config))
-        .add(new IssPcAccessConversionPass(config))
-        .add(new IssTcgContextPass(config))
-        .add(new IssSelectLoweringPass(config))
-        .add(new IssTcgSchedulingPass(config))
-        .add(new TcgBranchLoweringPass(config))
-        .add(new TcgOpLoweringPass(config))
-        .add(new IssHardcodedTcgAddOnPass(config))
-        .add(new IssTcgVAllocationPass(config))
-
-        // Common passes
-        .add(new IssGdbInfoExtractionPass(config))
-
-        // pre emit passes
-        .add(new IssCFunctionExtractionPass(config))
-        .add(new IssRegisterAccessInfoRetrievalPass(config))
-    ;
-
-    addDecodePasses(order, config);
-
-    if (config.dumpMode() == DumpMode.ISS_PASS_GRAPHS) {
-      addIssInstructionGraphDumpPasses(order, config);
-    }
-
-    addHtmlDump(order, config, "ISS Lowering Dump",
-        "This dump is executed after the iss transformation passes were executed.",
-        IssInfoRetrievalPass.class,
-        IssConfiguration.class);
-
-    if (!config.isDryRun()) {
-      // add iss template emitting passes to order
-      addIssEmitPasses(order, config);
-    }
-
-    return order;
+  public static PassOrder iss(IssConfiguration configuration) throws IOException {
+    return IssPassOrder.create(configuration);
   }
 
-  /**
-   * Adds {@link DumpIssInstructionGraphsPass} after each pass in ISS mode.
-   */
-  private static PassOrder addIssInstructionGraphDumpPasses(PassOrder order,
-                                                            GeneralConfiguration config) {
-    return order.addBetweenEach((current, next) -> {
-      if (current instanceof DumpIssInstructionGraphsPass) {
-        return Optional.empty();
-      }
-      return Optional.of(new DumpIssInstructionGraphsPass(config));
-    });
+  public static PassOrder rtl(RtlConfiguration configuration) throws IOException {
+    return RtlPassOrder.create(configuration);
   }
-
-  private static void addIssEmitPasses(PassOrder order, IssConfiguration config) {
-    order
-        // top-level meson build. just because we want to add target trace-events
-        .add(issDefault("meson.build", config))
-
-        // includes
-        .add(issDefault("/include/vadl-builtins.h", config))
-        .add(issDefault("/include/vadl-iss-builtins.h", config))
-
-        // config rendering
-        .add(issDefault("/configs/devices/gen-arch-softmmu/default.mak", config))
-        .add(issDefault("/configs/targets/gen-arch-softmmu.mak", config))
-        .add(issDefault("/target/gen-arch/cpu.h", config))
-
-        // arch init rendering
-        .add(issDefault("/include/disas/dis-asm.h", config))
-        .add(issDefault("/include/sysemu/arch_init.h", config))
-
-        // gdb rendering
-        // gdb-xml/gen-arch-cpu.xml
-        .add(new EmitIssGdbXmlPass(config))
-        // target/gen-arch/gdbsub.c
-        .add(new EmitIssGdbStubPass(config))
-
-        // plugin rendering
-        .add(issDefault("/tests/tcg/plugins/endstate.c", config))
-        .add(issDefault("/tests/tcg/plugins/meson.build", config))
-
-        // hardware rendering
-        .add(issDefault("/hw/Kconfig", config))
-        .add(issDefault("/hw/meson.build", config))
-        .add(issDefault("/hw/gen-arch/Kconfig", config))
-        .add(issDefault("/hw/gen-arch/meson.build", config))
-        .add(new EmitIssHwMachineCPass(config))
-        .add(issDefault("/hw/gen-arch/gen-machine.h", config))
-        .add(issDefault("/hw/gen-arch/boot.c", config))
-        .add(issDefault("/hw/gen-arch/boot.h", config))
-
-        // target rendering
-        .add(issDefault("/target/Kconfig", config))
-        .add(issDefault("/target/meson.build", config))
-        .add(issDefault("/target/gen-arch/trace-events", config))
-        .add(issDefault("/target/gen-arch/trace.h", config))
-        .add(issDefault("/target/gen-arch/Kconfig", config))
-        .add(issDefault("/target/gen-arch/meson.build", config))
-        .add(issDefault("/target/gen-arch/cpu-bits.h", config))
-        .add(new EmitIssHelperCPass(config))
-        // target/gen-arch/helper.h
-        .add(new EmitIssHelperHPass(config))
-        // target/gen-arch/do-exception.c.inc
-        .add(new EmitIssDoExcCIncPass(config))
-        // target/gen-arch/cpu-qom.h
-        .add(new EmitIssCpuQomHeaderPass(config))
-        // target/gen-arch/cpu-param.h
-        .add(new EmitIssCpuParamHeaderPass(config))
-        // target/gen-arch/cpu.h
-        .add(new EmitIssCpuHeaderPass(config))
-        // target/gen-arch/cpu.c
-        .add(new EmitIssCpuSourcePass(config))
-        // target/gen-arch/vdt-decode.c
-        .add(new EmitIssDecodeTreePass(config))
-        // target/gen-arch/insn_trans/trans_<isa>.c.inc
-        .add(new EmitIssInsnTransCIncPass(config))
-        // target/gen-arch/translate.c
-        .add(new EmitIssTranslateCPass(config))
-        // target/gen-arch/machine.c
-        .add(new EmitIssMachinePass(config))
-
-        // plugin rendering
-        .add(issDefault("/contrib/plugins/meson.build", config))
-        // cosimulation plugin
-        .add(issDefault("/contrib/plugins/cosimulation.c", config))
-    ;
-
-    addUserModeEmitPasses(order, config);
-  }
-
-  /**
-   * Adds all necessary passes for generating the VDT.
-   *
-   * @param order  into which the passes will be inserted.
-   * @param config from which to decide if a dump is wanted.
-   */
-  private static void addDecodePasses(PassOrder order, GeneralConfiguration config) {
-
-    if (config.getDecoderOptions().isDisabled(OPT_ALL)) {
-      return;
-    }
-
-    // VDT Decode Passes
-    order
-        .add(new VdtEncodingConstraintValidationPass(config))
-        .add(new VdtInputPreparationPass(config));
-
-    if (config.getDecoderOptions().isEnabled(OPT_CONSTRAINT_SYNTHESIS)) {
-      order.add(new VdtConstraintSynthesisPass(config));
-    }
-
-    if (config.getDecoderOptions().isEnabled(OPT_ENCODING_VERIFICATION)) {
-      order.add(new VdtEncodingSemanticVerificationPass(config));
-    }
-
-    order.add(new VdtLoweringPass(config));
-
-    if (config.getDecoderOptions().isEnabled(OPT_DECODER_VERIFICATION)) {
-      order.add(new VdtVerificationPass(config));
-    }
-  }
-
-  private static void addUserModeEmitPasses(PassOrder order, IssConfiguration config) {
-    // right now, we only emit those passes if we generate the ISS for RV64UME
-    var inputPath = config.inputPath();
-    if (inputPath != null && inputPath.getFileName().endsWith("rv64ume.vadl")) {
-      order
-          .add(issDefault("/configs/targets/gen-arch-linux-user.mak", config))
-
-          .add(issDefault("/linux-user/meson.build", config))
-          .add(issDefault("/linux-user/elfload.c", config))
-          .add(issDefault("/linux-user/syscall_defs.h", config))
-
-          .add(issDefault("/linux-user/gen-arch/meson.build", config))
-          .add(issDefault("/linux-user/gen-arch/cpu_loop.c", config))
-          .add(issDefault("/linux-user/gen-arch/signal.c", config))
-          .add(issDefault("/linux-user/gen-arch/sockbits.h", config))
-          .add(issDefault("/linux-user/gen-arch/syscall.tbl", config))
-          .add(issDefault("/linux-user/gen-arch/syscallhdr.sh", true, config))
-          .add(issDefault("/linux-user/gen-arch/target_cpu.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_elf.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_errno_defs.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_fcntl.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_mman.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_prctl.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_proc.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_resource.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_signal.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_structs.h", config))
-          .add(issDefault("/linux-user/gen-arch/target_syscall.h", config))
-          .add(issDefault("/linux-user/gen-arch/termbits.h", config))
-      ;
-    }
-  }
-
-  /**
-   * Constructs the pass order used to generate the RTL (Chisel) from a VADL specification.
-   */
-  public static PassOrder rtl(RtlConfiguration config) throws IOException {
-    var order = viam(config);
-
-    order.skip(NormalizeFieldsToFieldAccessFunctionsPass.class);
-    order.skip(RenamingConflictingRegistersPass.class);
-    order.skip(OverwriteInputOperandsPass.class);
-
-    order.add(new RtlConfigurationPass(config));
-
-    order.add(new StageOrderingPass(config));
-
-    order.add(new InstructionProgressGraphCreationPass(config))
-        .add(new MiaMappingCreationPass(config))
-        .add(new InstructionProgressGraphMergePass(config))
-        .add(new MiaMappingOptimizePass(config))
-        .add(new InstructionProgressGraphLowerPass(config))
-        .add(new InstructionProgressGraphNamePass(config));
-
-    addHtmlDump(order, config,
-        "mia-map",
-        "MiA after mapping instruction behavior");
-
-    order.add(new HazardAnalysisPass(config));
-
-    if (config.isEmitDebugPrint()) {
-      order.add(new DebugOutputPass(config));
-    }
-
-    order.add(new MiaMappingInlinePass(config));
-
-    if (config.isEmitRVFI()) {
-      order.add(new EmitRVFIOutputsPass(config));
-    }
-
-    order.add(new ForwardingLogicPass(config))
-        .add(new ControlLogicPass(config));
-
-    addHtmlDump(order, config,
-        "mia-inline",
-        "MiA after inlining instruction behavior");
-
-    if (config.getDecoderOptions().getGenerator() != RTL_TABLE) {
-      // Prepares and constructs the VDT, which is not used by the rtl-table strategy
-      addDecodePasses(order, config);
-    }
-
-    if (!config.isDryRun()) {
-      addRtlEmitPasses(order, config);
-    }
-
-    return order;
-  }
-
-  private static void addRtlEmitPasses(PassOrder order, RtlConfiguration config) {
-    order.add(new EmitBuildSbtPass(config))
-        .add(new EmitModulesPass(config))
-        .add(new EmitVadlLibPass(config))
-        // tests and simulation
-        .add(new EmitCoreTestPass(config))
-        .add(new EmitCoreEmitPass(config))
-        .add(new EmitElfSimPass(config))
-        .add(new EmitSimMemSvPass(config))
-        .add(new EmitSimMemCppPass(config))
-        // build files
-        .add(new EmitScalafmtConfigPass(config))
-        .add(new EmitRtlMakefilePass(config))
-        // devcontainer
-        .add(new EmitRtlDevcontainerConfigPass(config))
-        .add(new EmitRtlDevcontainerDockerComposePass(config))
-        .add(new CleanupEmitDirectoryPass(config));
-  }
-
-  /**
-   * Adds all necessary passes for a html dump of the VIAM if the config whishes to dump.
-   *
-   * @param order       into which the passes will be inserted.
-   * @param config      from which to decide if a dump is wanted.
-   * @param phase       is the name of the dump.
-   * @param description for the dump.
-   * @param exclusions  for which no behavior graph should be collected.
-   * @return the modified {@link PassOrder}.
-   */
-  private static PassOrder addHtmlDump(PassOrder order, GeneralConfiguration config,
-                                       String phase, String description, Class<?>... exclusions) {
-
-    if (config.doDump()) {
-      addDumpBehaviorCollectionPasses(order, config, exclusions);
-      var htmlConfig = HtmlDumpPass.Config.from(config, phase, description);
-      order.add(new HtmlDumpPass(htmlConfig));
-    }
-
-    return order;
-  }
-
-  /**
-   * Adds a {@link CollectBehaviorDotGraphPass} after each existing pass in the pass order.
-   * This function is idempotent, so calling it on the same pass order will not add
-   * more behavior collection passes.
-   *
-   * @param order      to add collection passes to
-   * @param exceptions pass classes that don't change the behavior and therefore should not have
-   *                   behavior collection pass successor.
-   * @return the passed order
-   */
-  private static PassOrder addDumpBehaviorCollectionPasses(PassOrder order,
-                                                           GeneralConfiguration config,
-                                                           Class<?>... exceptions) {
-    if (!config.doDump()) {
-      return order;
-    }
-
-    var dontRun = Streams.concat(Stream.of(
-        ViamVerificationPass.class,
-        AbstractTemplateRenderingPass.class,
-        CollectBehaviorDotGraphPass.class
-    ), Stream.of(exceptions)).toList();
-
-    order.addBetweenEach((prev, next) -> {
-
-      if (dontRun.stream().anyMatch(a -> a.isInstance(prev))) {
-        // if previous is any of the dont run pass classes, don't add the collection behavior
-        return Optional.empty();
-      }
-
-      if (next.orElse(null) instanceof CollectBehaviorDotGraphPass) {
-        // don't add a collection pass, if the next one is already a collection pass.
-        return Optional.empty();
-      }
-
-      return Optional.of(new CollectBehaviorDotGraphPass(config));
-    });
-    return order;
-  }
-
 
   /**
    * A pseudo pass that indicates the first pass in the PassOrder.
-   * It is necessary to dump the behavior directly after creation,
-   * before any other pass manipulated the behavior.
-   *
-   * <p>This pass contains no logic.</p>
    */
   public static class ViamCreationPass extends Pass {
-
     public ViamCreationPass(GeneralConfiguration configuration) {
       super(configuration);
     }
@@ -892,5 +78,4 @@ public class PassOrders {
       return null;
     }
   }
-
 }
